pyproject.toml
[project]
name = "art-agent"
version = "0.1.0"
description = "Autonomous pixel art evolution via PixelGPT + genetic selection"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "textual>=3.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
